requirements.txt
Flask
deepgram-sdk
twilio
python-dotenv
openai
sounddevice
langchain
langchain-groq
langchain-chroma
langchain-community
langchain-openai
PyDub
ffmpeg
Flask-Session
langchain-pinecone
pinecone
pypdf
unstructured
langchain-together